datus-oracle/pyproject.toml
[project]
name = "datus-oracle"
version = "0.1.0"
description = "Oracle database adapter for Datus using python-oracledb"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
authors = [
    {name = "DatusAI", email = "support@datus.ai"}
]
keywords = ["datus", "database", "oracle", "adapter"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "datus-agent>0.2.1",
    "datus-sqlalchemy>=0.1.0",
    "oracledb>=3.4.0",
    "pydantic>=2.5.0",
]

[project.urls]
Homepage = "https://github.com/Datus-ai/datus-db-adapters"
Repository = "https://github.com/Datus-ai/datus-db-adapters"
Issues = "https://github.com/Datus-ai/datus-db-adapters/issues"

[project.entry-points."datus.adapters"]
oracle = "datus_oracle:register"

[tool.uv.sources]
datus-sqlalchemy = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["datus_oracle"]
